+++ src/components/common/SectionTitle.js
@@ -21,13 +21,13 @@ const SectionTitle = ({ classes, title, ...props }) => {
 
     return (
         <Grid container alignItems="center">
-            <Grid className={internalClasses.line} item xs={4}></Grid>
+            <Grid className={clsx(internalClasses.line, classes?.line)} item xs={4}></Grid>
             <Grid item xs={4}>
-                <Typography variant="h1" className={clsx(internalClasses.title, 'fs-300 fw-500')}>
+                <Typography variant="h1" className={clsx(internalClasses.title, classes?.title, 'fs-300 fw-500')}>
                     {title}
                 </Typography>
             </Grid>
-            <Grid className={internalClasses.line} item xs={4}></Grid>
+            <Grid className={clsx(internalClasses.line, classes?.line)} item xs={4}></Grid>
         </Grid>
     );
 };