+++ src/data/webDesign.js
@@ -1,0 +1,52 @@
+const array = [
+    {
+        id: 'tierra-roja',
+        website: 'https://tierra-roja.herokuapp.com/',
+        title: 'Tierra roja',
+        en: {
+            subtitle: 'Web app for clay artists in Misiones.'
+        },
+        es: {
+            subtitle: 'Web app for clay artists in Misiones.'
+        },
+        type: 'web-design'
+    },
+    {
+        id: 'chapati',
+        website: 'https://www.chapati-delta.com.ar/',
+        title: 'Chapati',
+        en: {
+            subtitle: "Conscious website promoting the Chapati' project"
+        },
+        es: {
+            subtitle: "Conscious website promoting the Chapati' project"
+        },
+        type: 'web-design'
+    },
+    {
+        id: 'tomi-scattini',
+        title: 'Tomi Scattini',
+        website: 'https://tomiscattini.com/',
+        en: {
+            subtitle: 'Full stack web developer Portfolio'
+        },
+        es: {
+            subtitle: 'Full stack web developer Portfolio'
+        },
+        type: 'web-design'
+    },
+    {
+        id: 'vaikuntha',
+        website: 'https://vaikuntha.com.ar/',
+        title: 'Vaikuntha',
+        en: {
+            subtitle: 'Coming soon!'
+        },
+        es: {
+            subtitle: 'Coming soon!'
+        },
+        type: 'web-design'
+    }
+];
+
+export default array;